Xenide/requirements.txt
discord.py
aiohttp
python-dotenv
Pillow
numpy